+++ omeroweb/feedback/urls.py
@@ -23,17 +23,19 @@
 # Version: 1.0
 #
 
-from django.conf.urls import url
+from django.urls import re_path
 from django.views.generic import TemplateView
 
 from omeroweb.feedback import views
 
 # url patterns
 urlpatterns = [
-    url(r"^feedback/", views.send_feedback, name="fsend"),
-    url(r"^comment/", views.send_comment, name="csend"),
-    url(r"^thanks/", TemplateView.as_view(template_name="thanks.html"), name="fthanks"),
-    url(
+    re_path(r"^feedback/", views.send_feedback, name="fsend"),
+    re_path(r"^comment/", views.send_comment, name="csend"),
+    re_path(
+        r"^thanks/", TemplateView.as_view(template_name="thanks.html"), name="fthanks"
+    ),
+    re_path(
         r"^disabled/",
         TemplateView.as_view(template_name="disabled.html"),
         name="feedback_disabled",