+++ omeroweb/webgateway/admin.py
@@ -15,7 +15,7 @@
 
 # from models import StoredConnection
 # from django.contrib import admin
-# from django.utils.translation import ugettext_lazy as _
+# from django.utils.translation import gettext_lazy as _
 
 # class StoredConnectionOptions(admin.ModelAdmin):
 #  list_display = ('base_path', 'config_file', 'username', 'failcount')